# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: Help API
  version: "1.0"
servers:
  - url: http://192.168.25.107:8081
    description: Generated server url
paths:
...
  /documents/{pageId}:
    get:
      tags:
        - stock-document-controller
      operationId: getDocuments
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageStockDocument'
  /documents/document/{docId}:
    get:
      tags:
        - stock-document-controller
      operationId: getDocument
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockDocument'
  /documents:
    post:
      tags:
        - stock-document-controller
      operationId: addDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockDocument'
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockDocument'
  /inventory/{docId}:
    post:
      tags:
        - stock-inventory-controller
      operationId: addInventories
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StockInventory'
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockDocument'
  /items/item/{id}:
    get:
      tags:
        - stock-item-balance-controller
      operationId: getItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockItemBalance'
    post:
      tags:
        - stock-item-balance-controller
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockItemBalance'
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockItemBalance'
  /items:
    get:
      tags:
        - stock-item-balance-controller
      operationId: getAllItems
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: category
          in: query
          required: true
          schema:
            type: string
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItemBalance'
    post:
      tags:
        - stock-item-balance-controller
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockItemBalance'
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StockItemBalance'
  /items/delete/{id}:
    post:
      tags:
        - stock-item-balance-controller
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "409":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "400":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "404":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "406":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "403":
          description: default response
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        "200":
          description: default response
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int32
        category:
          type: string
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    Units:
      type: object
      properties:
        id:
          type: integer
          format: int32
        unitName:
          type: string
...
    PageStockDocument:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/StockDocument'
        empty:
          type: boolean
    StockDocument:
      type: object
      properties:
        documentId:
          type: integer
          format: int64
        documentType:
          type: string
        date:
          type: string
          format: date-time
        inventories:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/StockInventory'
    StockInventory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        stockDocument:
          $ref: '#/components/schemas/StockDocument'
        productId:
          type: integer
          format: int64
        productName:
          type: string
        amount:
          type: number
          format: double
    StockItemBalance:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        unit:
          $ref: '#/components/schemas/Units'
        category:
          $ref: '#/components/schemas/Category'
        inStockQty:
          type: number
          format: double
        productId:
          type: integer
          format: int64
        restaurant:
          type: boolean
...